Organize agsys into 3 subdirectories
Based on discussion with Bin Peng and Erik Kluzek.

The idea is:

The science directory should be portable to different land models. Code
there should only depend on other code in the science directory plus
code in the ctsm_wrappers directory.

The ctsm_wrappers directory will provide wrappers to constants,
variables and low-level routines in CTSM that are used by the AgSys
science code in order to make it operate consistently with the rest of
CTSM. If AgSys were plugged in to a different land model, the code in
this directory would need to be replaced by alternatives that provided
the same interface but referenced constants, variables and routines that
are available in the given host model.

The ctsm_interface directory will contain code that is specific to the
CTSM implementation of AgSys. Furthermore, CTSM should only interact
with AgSys via the modules here.