Rework phases and params derived types
Rather than having one instance that stores information for all crop
types, instead have a given instance store information only for a single
crop type, with an array of instances in the interface layer (one for
each crop type). Bin Peng suggested this to simplify the science code:
now, when calling a science routine for a given patch, the interface
layer can pass in just the phases and params instances needed for the
crop (and cultivar) type contained in that patch.

Also, add a derived type for storing information about composite phases.

Also: shoot_lag and shoot_rate are now crop-specific rather than
cultivar-specific parameters (according to Bin Peng).

Also edit the documentation of the arrays in agsys_phases_type: I found
it confusing to think of these as indexed by stage, and find it more
intuitive to think of these as indexed by phase. (But I'm not sure if
this is actually a correct way to think about these arrays....)