Pack environmental inputs into a derived type
This will simplify the interface to many of the subroutines in the
science code. This is especially helpful with the migration of the
science code to being more object-oriented: routines can accept an
instance of this environmental inputs type rather than needing to accept
all of the possible inputs needed by any possible child classes.

We considered doing the same thing with the inout arguments to the
phenology routine. We may decide to do that in the future, but for now
we're not because:
- There seems to be less benefit from doing this (in terms of
  simplifying argument lists)
- It can be nice to have the outputs explicitly listed
- This would require two calls rather than one in the interface layer:
  one to set the old values of each state variable, and one to get the
  new values after the phenology call (so there is more overhead in
  terms of code and performance)