Switch dimension ordering for 2-d AgSys variables
This seems to be needed for the sake of history routines. Here is a
relevant todo note that I just added to the code:

TODO(wjs, 2019-11-13) 2-d variables are stored with patch (or column) as
the first dimension. This follows CTSM conventions, but is less
efficient for AgSys, where we operate on a single point at a
time. Ideally, for the sake of performance, we would switch this
dimension order for 2-d AgSys variables. However, currently that
prevents us from doing history output for the given 2-d variable. (Also,
although the restart routines seem to be general enough to handle
variables with either dimension ordering - via switchdim - it doesn't
look like there are any variables that currently have switchdim =
.false., so some changes may be needed to support that robustly, in the
restart utilities and/or in init_interp.) At some point, we should
consider putting in place the necessary generalization for the history
infrastructure to handle 2-d variables with this switched dimension
ordering (as long as this can be done in a way that doesn't hurt the
efficiency of the history infrastructure). At that point, we should
change the variables here so that patch (or column) is the second
dimension.