fix a bug related with gddplant_patch
The bug is that gddplant_patch is modified outside the accumulator. In the subroutine CropPhenology,
a maximum-lai-trigger for transition to grain filling phase was implemented by directly change the value of hui (gddplant_patch).
This leads to a sudden jump in the gddplant, which is the accumulated gdd since sowing for crops.